refactor parse module
